+++ GameTanks/base_level.cpp
@@ -127,7 +127,7 @@ void BaseLevel::SetBorderTexture(std::string texture_address, int const& size_le
 	size_level_border_ = size_level_border;
 }
 
-void BaseLevel::SetBackgroundMusic(std::string music_address, int const& volume) {
+void BaseLevel::SetBackgroundMusic(std::string music_address, float const& volume) {
 	music_background_.openFromFile(music_address);
 	music_background_.play();
 	music_background_.setVolume(volume);
